% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{analyseCorrelationCoefficient}
\alias{analyseCorrelationCoefficient}
\title{Analyse correlation coefficient}
\usage{
analyseCorrelationCoefficient(d, i)
}
\description{
Analyse correlation coefficient
}
\examples{
analyseCorrelationCoefficient()
}
\keyword{gpclips}
