% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1.R
\name{analyzePercentJapan}
\alias{analyzePercentJapan}
\title{Analyse percentage Japan}
\usage{
analyzePercentJapan(d, i)
}
\description{
Analyse percentage Japan
}
\examples{
analyzePercentJapan()
}
\keyword{gpclips}
