% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1.R
\name{calcAbsJapan}
\alias{calcAbsJapan}
\title{Bootstrap method to calculate abs female going to Japan}
\usage{
calcAbsJapan(x, index)
}
\description{
Bootstrap method to calculate abs female going to Japan
}
\examples{
calcAbsJapan()
}
\keyword{gpclips}
