% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{calcanalyseCorrelationCoefficient}
\alias{calcanalyseCorrelationCoefficient}
\title{Bootstrap method to calculate correlation coefficient}
\usage{
calcanalyseCorrelationCoefficient(x, index)
}
\description{
Bootstrap method to calculate correlation coefficient
}
\examples{
calcanalyseCorrelationCoefficient()
}
\keyword{gpclips}
