% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{calcanalyseCovariance}
\alias{calcanalyseCovariance}
\title{Bootstrap method to calculate covariance}
\usage{
calcanalyseCovariance(x, index)
}
\description{
Bootstrap method to calculate covariance
}
\examples{
calcanalyseCovariance()
}
\keyword{gpclips}
