% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1.R
\name{calcPercentJapan}
\alias{calcPercentJapan}
\title{Bootstrap method to calculate percentage female going to Japan}
\usage{
calcPercentJapan(x, index)
}
\description{
Bootstrap method to calculate percentage female going to Japan
}
\examples{
calcPercentJapan()
}
\keyword{gpclips}
