% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{init_projekt}
\alias{init_projekt}
\title{Inits the project by loading the required resources}
\usage{
init_projekt()
}
\description{
Inits the project by loading the required resources
}
\examples{
init_projekt()
}
\keyword{init_projekt}
