% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plotLm}
\alias{plotLm}
\title{Plot LM using ggplot}
\usage{
plotLm(xData, xLab, yLab, title)
}
\arguments{
\item{xData, }{xLab, yLab, title}
}
\description{
Plot LM using ggplot
}
\examples{
plotLm()
}
\keyword{gpclips}
