% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_1.R
\name{prepareDataH1}
\alias{prepareDataH1}
\title{Returns data prepared for hypothesis 1}
\usage{
prepareDataH1()
}
\description{
Returns data prepared for hypothesis 1
}
\examples{
prepareDataH1()
}
\keyword{gpclips}
