% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{prepareDataH2}
\alias{prepareDataH2}
\title{Returns data prepared for hypothesis 1}
\usage{
prepareDataH2()
}
\description{
Returns data prepared for hypothesis 1
}
\examples{
prepareDataH2()
}
\keyword{gpclips}
