% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MellinQF.R
\name{print.MellinQF}
\alias{print.MellinQF}
\title{Printing \code{MellinQF}}
\usage{
\method{print}{MellinQF}(x, digits = 3L, ...)
}
\arguments{
\item{x}{\code{MellinQF} object.}

\item{digits}{number of digits to display in the output.}

\item{...}{potentially more arguments passed to methods.}
}
\description{
It allows to visualize useful information on the \code{MellinQF} object.
}
