% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MellinQF_ratio.R
\name{print.MellinQF_ratio}
\alias{print.MellinQF_ratio}
\title{Print \code{MellinQF_ratio}}
\usage{
\method{print}{MellinQF_ratio}(x, digits = 3L, ...)
}
\arguments{
\item{x}{\code{MellinQF_ratio} object.}

\item{digits}{number of digits to display in the output.}

\item{...}{potentially more arguments passed to methods.}
}
\description{
It allows to visualize useful information on the \code{MellinQF_ratio} object.
}
