% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_3.R
\name{analyzePercentJapan}
\alias{analyzePercentJapan}
\title{Analyse percent Japan}
\usage{
analyzePercentJapan(d, i)
}
\description{
Analyse percent Japan
}
\examples{
analyzePercentJapan()
}
\keyword{gpclips}
